Add AwaEmbedding and LLMEmbedding for users
