+++ awadb/llm_embedding/__init__.py
@@ -5,18 +5,30 @@ from sentence_transformers import SentenceTransformer
 from transformers import AutoModel
 from transformers import AutoTokenizer
 from typing import Iterable, Any, List
-from awadb import llm_embedding
 
 class LLMEmbedding:
-    """Embedding models."""
-    def __init__(self, model_name):
-        self.model_name = model_name
-        if self.model_name == "HuggingFace":
-            from awadb.llm_embedding.huggingface import HuggingFaceEmbeddings
-            self.llm = HuggingFaceEmbeddings()
-        elif self.model_name == "OpenAI":
-            from awadb.llm_embedding.openai import OpenAIEmbeddings
-            self.llm = OpenAIEmbeddings()
+    def __init__(self):
+        self.model = SentenceTransformer('all-MiniLM-L6-v2')
+        self.tokenizer = None 
+
+
+    def Embedding(self, sentence):
+        tokens = []
+        if self.tokenizer != None:
+            tokens = self.tokenizer.tokenize(sentence)
+        else:
+            tokens.append(sentence)
+        return self.model.encode(tokens[0])
+
+    def EmbeddingBatch(
+        self,
+        texts: Iterable[str],
+        **kwargs: Any,
+    ) -> List[List[float]]:
+        results: List[List[float]] = []
+        for text in texts:
+            results.append(self.model.encode(text))
+        return results
 
     #set your own llm
     def SetModel(self, model_name):
@@ -26,12 +38,5 @@ class LLMEmbedding:
     def SetTokenizer(self, tokenizer_name):
         self.tokenizer = AutoTokenizer.from_pretrained(tokenizer_name)
 
-    def Embedding(self, sentence):
-        return self.llm.Embedding(sentence)
-    
-    def EmbeddingBatch(
-        self,
-        texts: Iterable[str],
-        **kwargs: Any,
-    ) -> List[List[float]]:
-        return self.llm.EmbeddingBatch(texts, **kwargs)
+
+
